added sessions and improved auth
