+++ server/package.json
@@ -15,6 +15,7 @@
     "connect-flash": "^0.1.1",
     "ejs": "^3.1.6",
     "express": "^4.17.1",
+    "express-session": "^1.17.2",
     "faker": "^5.5.3",
     "mysql": "^2.18.1",
     "nodemon": "^2.0.12"
