+++ web/views/pages/appointment.ejs
@@ -9,7 +9,7 @@
     <div class="col-6 image-part"></div>
 
     <div class="apt col-6 reg d-flex flex-column">
-      <img src="./images/toplogo.png" alt="logo" class="logo" />
+      <img src="/images/toplogo.png" alt="logo" class="logo" />
       <div class="login-title text-center">Appointment</div>
       <form method = "POST" action="/appointment">
         <div class="row">
