+++ web/views/pages/home.ejs
@@ -22,7 +22,7 @@
         </div>
         <div class="col-sm-2 d-none d-lg-block appoint">
           <button class="btn btn-success">
-            <a id="btxt" href="patient/register">Book an Appointment </a>
+            <a id="btxt" href="patient/appointment">Book an Appointment </a>
           </button>
         </div>
       </div>
@@ -66,7 +66,7 @@
 
           <div class="row vbh">
             <div class="btn btn-default animated bounceInUp">
-              <a href="/patient/register">Book an Appointment </a>
+              <a href="/patient/appointment">Book an Appointment </a>
             </div>
           </div>
         </div>
@@ -90,7 +90,7 @@
 
           <div class="row vbh">
             <div class="btn btn-default animated bounceInUp">
-              <a href="/patient/register">Book an Appointment </a>
+              <a href="/patient/appointment">Book an Appointment </a>
             </div>
           </div>
         </div>
