+++ web/views/pages/patientdash.ejs
@@ -348,7 +348,7 @@
                             <button class="btn btn-secondary" type="button" data-dismiss="modal">
                                 Cancel
                             </button>
-                            <a class="btn btn-primary" href="/patient/login">Logout</a>
+                            <form action="/patient/logout" method="POST"><button class="btn btn-primary">Logout</button></form>
                         </div>
                     </div>
                 </div>
